+++ config/toml.go
@@ -486,6 +486,10 @@ chunk_request_timeout = "{{ .StateSync.ChunkRequestTimeout }}"
 # The number of concurrent chunk fetchers to run (default: 1).
 chunk_fetchers = "{{ .StateSync.ChunkFetchers }}"
 
+# Maximum number of chunks allowed in a snapshot (default: 100000).
+# If set to 0, the default will be used.
+max_snapshot_chunks = {{ .StateSync.MaxSnapshotChunks }}
+
 #######################################################
 ###       Block Sync Configuration Options          ###
 #######################################################