+++ statesync/messages_test.go
@@ -69,12 +69,17 @@ func TestValidateMsg(t *testing.T) {
 			false},
 		"SnapshotsResponse no hash": {
 			&ssproto.SnapshotsResponse{Height: 1, Format: 1, Chunks: 2, Hash: []byte{}},
-			false},
+			false,
+		},
+		"SnapshotsResponse exceeds max chunks": {
+			&ssproto.SnapshotsResponse{Height: 1, Format: 1, Chunks: 100001, Hash: []byte{1}},
+			false,
+		},
 	}
 	for name, tc := range testcases {
 		tc := tc
 		t.Run(name, func(t *testing.T) {
-			err := validateMsg(tc.msg)
+			err := validateMsg(tc.msg, 100000)
 			if tc.valid {
 				require.NoError(t, err)
 			} else {